This is a waypoint version used to demonstrate that the RCS calculations now work for dihedrals and multiple bounces. The issue in the previous version was caused by calculating the power after each bounce. When two hit points were very close together then the inverse square law was increasing the reflected power. To fix this, the inverse square calculation is now only performed once just before the PO calculation. This means that the total range is now used to determine the power which removes this problem.
Also in order to make the PO calculations work for dihedrals the range to the receiver has to be the correct range for each ray from its reflection point. For normal PO calculations, the range from the origin is used. This means that the phase of each of the PO calculations for each triangle have a common reference point and can be coherently combined together. (i.e. two triangles forming a square provide the equivalent beam function of a square - a sinc function). When we calculate the PO for only the last bounce then in order to make dihedrals (and trihedrals) form correctly, the total phase after all bounces must be identical. This means that the hotpoint of the final ray has to be used as the reference. An idea to fix this in the final version is to use the origin for first bounce PO calculations and the ray-origin for higher bounces.