fastKDTree now works. Its not optimised yet but it performs the Zhou algorithm. It currently saves the kdtree in the same file format as the older buildKdTree code (which was written in Objective-C) for debugging purposes but this needs changing as the new code save the tree in an optimised format thereby saving conversion time in sarcastic.
Also
BUG FIX: this needs incorporating into Sarcastic but remains in this branch for now. The accelerated triangle structure was saving the dimension parameter k as a double in the file. Fortunately it was consistent throughout sarcastic but it could potentially cause problems as its used to index arrays. its been fixed in this code both in the buildKdTree code and the readKdTree code.