This version has a new improved version of materialise. WARNING: All previous versions of materialise have a bug in them. There are two main big fixes here: 1) The previous version of materialise did not correctly grow a region around a 'seed' triangle from the input mesh. This meant that sometimes the input to the triangulation function could have two distinct polygons with individual bounding segments. This could cause crashes and hangs in the triangulation function 2) The previous versions made no distinction between a triangle mesh and a triangle soup. The output form Sketchup and other common programmes is more usually a 'triangle soup' where the triangles are just saved to file with their vertex information. In these cases a vertex may only belong to one triangle and might sit part way along the side of a neighboring triangle. This means that the subject triangle can have an edge that is adjoining a triangle and also has no neighbor. This version of materialise searches for these 'polygamous' triangle sides (with more than one partner ;-) ) and makes them monogamous by spitting them at the offending vertex. By doing this it converts the triangle soup into a 'mesh'. It was intended to make this function run quicker and in doing so I used the CGAL (Computational Geometry Algorithms) library instead of Jon Shewchuk's Triangle library. I suspect that CGAL is slower than 'Triangle' but the latter has terrible exception control and was impossible to debug. In fact I would never have been able to find the above problems using Jons library as it would just hang or spin around strange exceptions when the above conditions were met. I may revisit Triangle in the future.
It is worth noting that there is still a unhandled problem in this version. Sometimes a triangle in the input soup can have vertices that are completely inscribed inside another triangle . As I only check for polygamous 'sides' rather than sides that cross each other then an error can still occur. This doesn't occur often and can be 'managed' by carefully building your model. For now if the code hangs or crashes load it into a debugger and find the value of 'cnt' in the main function. Then use this to set the DEBUGCOUNT in main.cpp and run it again. This time it will save the offending mesh, points, triangles and segments out so that you can fix your model.